requirements.txt
httpie==0.9.2
psycopg2==2.6
pytest==2.6.4
pyjade==3.0.0
SQLAlchemy==0.9.8
factory-boy==2.4.1
fake-factory==0.4.2
flake8==2.3.0
Flask==0.10.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
